requirements.txt
appdirs==1.4.3
asn1crypto==0.22.0
cffi==1.10.0
cryptography==1.9
Django==1.10.6
djangorestframework==3.6.3
elasticsearch==5.2.0
-e git+git@github.com:mattfield11/python-opcua.git@479054f84624653f60924d07d80ce6ecf8ca24d8#egg=freeopcua
future==0.16.0
idna==2.5
influxdb==4.0.0
ldap3==2.2.2
lxml==3.8.0
packaging==16.8
paho-mqtt==1.2.2
pika==0.10.0
pkg-resources==0.0.0
protobuf==3.2.0
psycopg2==2.7.1
pyasn1==0.2.3
pycparser==2.17
pyparsing==2.2.0
python-dateutil==2.6.0
pytz==2016.10
requests==2.13.0
six==1.10.0
urllib3==1.20
uWSGI==2.0.14
